Move Types out of the undertow package.
